dded: Autorun media in DVD drive (Thx 2 bobbin007)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@178 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/mplayer/xbox_video.cpp
@@ -282,12 +282,12 @@ static unsigned int Directx_ManageDisplay(unsigned int width,unsigned int height
 	if( !(g_graphicsContext.IsFullScreenVideo() || g_graphicsContext.IsCalibrating() ))
 	{
 		const RECT& rv = g_graphicsContext.GetViewWindow();
-		iScreenWidth = rv.right-rv.left;
-		iScreenHeight= rv.bottom-rv.top;
-		fOffsetX1    = rv.left;
-		fOffsetY1    = rv.top;
-		fOffsetX2    = 0;
-		fOffsetY2    = 0;
+		iScreenWidth = (float)rv.right-rv.left;
+		iScreenHeight= (float)rv.bottom-rv.top;
+		fOffsetX1    = (float)rv.left;
+		fOffsetY1    = (float)rv.top;
+		fOffsetX2    = 0.0f;
+		fOffsetY2    = 0.0f;
 	}
 	//if( g_graphicsContext.IsFullScreenVideo() || g_graphicsContext.IsCalibrating() )
   {
